backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python
pillow
python-multipart
